Move Output writing to separate module
