+++ generate/src/write.rs
@@ -1,0 +1,107 @@
+use crate::output::Output;
+use crate::{is_xid_continue, is_xid_start, CHUNK};
+
+pub fn output(index_start: &[u8], index_continue: &[u8], halfdense: &[u8]) -> Output {
+    let mut out = Output::new();
+    writeln!(
+        out,
+        "// \x40generated by ../generate. To regenerate, run the following in the repo root:",
+    );
+    writeln!(out, "//");
+    writeln!(
+        out,
+        "// $ curl -LO https://www.unicode.org/Public/zipped/14.0.0/UCD.zip",
+    );
+    writeln!(out, "// $ unzip UCD.zip -d UCD");
+    writeln!(out, "// $ cargo run --manifest-path generate/Cargo.toml");
+    writeln!(out);
+
+    writeln!(out, "const T: bool = true;");
+    writeln!(out, "const F: bool = false;");
+    writeln!(out);
+
+    writeln!(out, "#[repr(C, align(8))]");
+    writeln!(out, "pub(crate) struct Align8<T>(pub(crate) T);");
+    writeln!(out, "#[repr(C, align(64))]");
+    writeln!(out, "pub(crate) struct Align64<T>(pub(crate) T);");
+    writeln!(out);
+
+    writeln!(
+        out,
+        "pub(crate) static ASCII_START: Align64<[bool; 128]> = Align64([",
+    );
+    for i in 0u8..4 {
+        write!(out, "   ");
+        for j in 0..32 {
+            let ch = (i * 32 + j) as char;
+            write!(out, " {},", if is_xid_start(ch) { 'T' } else { 'F' });
+        }
+        writeln!(out);
+    }
+    writeln!(out, "]);");
+    writeln!(out);
+
+    writeln!(
+        out,
+        "pub(crate) static ASCII_CONTINUE: Align64<[bool; 128]> = Align64([",
+    );
+    for i in 0u8..4 {
+        write!(out, "   ");
+        for j in 0..32 {
+            let ch = (i * 32 + j) as char;
+            write!(out, " {},", if is_xid_continue(ch) { 'T' } else { 'F' });
+        }
+        writeln!(out);
+    }
+    writeln!(out, "]);");
+    writeln!(out);
+
+    writeln!(out, "pub(crate) const CHUNK: usize = {};", CHUNK);
+    writeln!(out);
+
+    writeln!(
+        out,
+        "pub(crate) static TRIE_START: Align8<[u8; {}]> = Align8([",
+        index_start.len(),
+    );
+    for line in index_start.chunks(16) {
+        write!(out, "   ");
+        for byte in line {
+            write!(out, " 0x{:02X},", byte);
+        }
+        writeln!(out);
+    }
+    writeln!(out, "]);");
+    writeln!(out);
+
+    writeln!(
+        out,
+        "pub(crate) static TRIE_CONTINUE: Align8<[u8; {}]> = Align8([",
+        index_continue.len(),
+    );
+    for line in index_continue.chunks(16) {
+        write!(out, "   ");
+        for byte in line {
+            write!(out, " 0x{:02X},", byte);
+        }
+        writeln!(out);
+    }
+    writeln!(out, "]);");
+    writeln!(out);
+
+    writeln!(
+        out,
+        "pub(crate) static LEAF: Align64<[u8; {}]> = Align64([",
+        halfdense.len(),
+    );
+    for line in halfdense.chunks(16) {
+        write!(out, "   ");
+        for byte in line {
+            write!(out, " 0x{:02X},", byte);
+        }
+        writeln!(out);
+    }
+    writeln!(out, "]);");
+
+    out
+}
